dedShop/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-autoslug==1.9.8
pytz==2021.1
sqlparse==0.4.2
